requirements.txt
av
pillow
pyside6
pyee
folium